requirements.txt
openai 
langchain
langchain-openai
langchain_pinecone
langchain[docarray]
docarray
pydantic==1.10.16
pytube 
pytubefix
python-dotenv
tiktoken 
pinecone-client 
scikit-learn
ruff
git+https://github.com/openai/whisper.git